continuing iteration for the instant ride flow

+++ lib/app/menu/views/scheduled_trips_menu/views/scheduled_rides_menu/content/cancelled_scheduled_ride_container.dart
@@ -12,7 +12,7 @@ cancelledScheduledRideContainer(
   String? driverName,
   String? vehicleName,
   String? pickup,
-  int? amount,
+  double? amount,
   String? destination,
 }) {
   return Column(
