+++ lib/app/menu/views/scheduled_trips_menu/views/scheduled_rides_menu/content/pending_scheduled_ride_container.dart
@@ -12,7 +12,7 @@ pendingScheduledRideContainer(
   String? driverName,
   String? vehicleName,
   String? pickup,
-  int? amount,
+  double? amount,
   String? destination,
   void Function()? viewPendingScheduledRide,
 }) {
